refactor(static-slicing): move static slicing steps accordingly
