+++ src/core/steps/steps.ts
@@ -37,8 +37,8 @@ import { IStep, StepHasToBeExecuted } from './step'
 import { PARSE_WITH_R_SHELL_STEP } from './all/00-parse'
 import { NORMALIZE } from './all/10-normalize'
 import { LEGACY_STATIC_DATAFLOW } from './all/20-dataflow'
-import { STATIC_SLICE } from './all/30-slice'
-import { NAIVE_RECONSTRUCT } from './all/40-reconstruct'
+import { STATIC_SLICE } from './all/static-slicing/30-slice'
+import { NAIVE_RECONSTRUCT } from './all/static-slicing/40-reconstruct'
 
 
 export const STEPS_PER_FILE = {
