+++ test/functionality/_helper/shell.ts
@@ -19,7 +19,7 @@ import { SlicingCriteria } from '../../../src/slicing'
 import { testRequiresRVersion } from './version'
 import { deepMergeObject, MergeableRecord } from '../../../src/util/objects'
 import { LAST_STEP, SteppingSlicer } from '../../../src/core'
-import { NAIVE_RECONSTRUCT } from '../../../src/core/steps/all/40-reconstruct'
+import { NAIVE_RECONSTRUCT } from '../../../src/core/steps/all/static-slicing/40-reconstruct'
 
 export const testWithShell = (msg: string, fn: (shell: RShell, test: Mocha.Context) => void | Promise<void>): Mocha.Test => {
 	return it(msg, async function(): Promise<void> {
